Correcciones de PR
